requirements/base.txt
# =========================
# Core Django
# =========================
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=24.1
django-cors-headers>=4.3
pytz>=2024.1

# =========================
# Auth / Security
# =========================
django-axes>=6.3
django-otp>=1.2
pyotp>=2.8
cryptography>=42.0.0
django-encrypted-model-fields>=0.6.5
qrcode>=7.4

# =========================
# Database
# =========================
psycopg2-binary>=2.9

# =========================
# Async / WebSockets
# =========================
daphne>=4.0
channels>=4.0
channels_redis>=4.0
uvicorn>=0.27

# =========================
# Celery / Background Tasks
# =========================
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
django-celery-results>=2.5

# =========================
# API / Schema
# =========================
drf-spectacular>=0.27

# =========================
# Caching
# =========================
django-redis>=5.4

# =========================
# Storage
# =========================
boto3>=1.34
django-storages>=1.14
whitenoise>=6.5
Pillow>=10.2
python-magic>=0.4

# =========================
# Search
# =========================
#django-elasticsearch-dsl>=8.0

# =========================
# AI / ML (used by ai_services)
# =========================
#openai>=1.10
#langchain>=0.1
#sentence-transformers>=2.3
#spacy>=3.7

# =========================
# Reporting / Data
# =========================
pandas>=2.2
openpyxl>=3.1
reportlab>=4.1
WeasyPrint>=61.0

# =========================
# Utils
# =========================
python-dateutil>=2.8
python-decouple>=3.8
requests>=2.31
python-json-logger>=2.0.7

# =========================
# Payments
# =========================
razorpay>=1.4

# =========================
# Monitoring
# =========================
sentry-sdk>=1.40
prometheus-client>=0.19

# =========================
# GraphQL
# =========================
strawberry-graphql-django>=0.38
